Separate the implementation of the `SpatialJoin` operation into a separate class (#1552)
This makes the code cleaner by separating the boilerplate required as a  subclass of `Operation` from the actual geometric calculations. It also makes the development much more convenient, as changing details in the implementation doesn't require the recompilation of the complete `engine` library.
