+++ src/engine/SpatialJoinAlgorithms.cpp
@@ -1,0 +1,225 @@
+#include "engine/SpatialJoinAlgorithms.h"
+
+#include <s2/s2closest_point_query.h>
+#include <s2/s2earth.h>
+#include <s2/s2point.h>
+#include <s2/s2point_index.h>
+#include <s2/util/units/length-units.h>
+
+#include "util/GeoSparqlHelpers.h"
+
+// ____________________________________________________________________________
+SpatialJoinAlgorithms::SpatialJoinAlgorithms(
+    QueryExecutionContext* qec, PreparedSpatialJoinParams params,
+    bool addDistToResult,
+    std::variant<NearestNeighborsConfig, MaxDistanceConfig> config)
+    : qec_{qec},
+      params_{std::move(params)},
+      addDistToResult_{addDistToResult},
+      config_{std::move(config)} {}
+
+// ____________________________________________________________________________
+std::optional<GeoPoint> SpatialJoinAlgorithms::getPoint(const IdTable* restable,
+                                                        size_t row,
+                                                        ColumnIndex col) const {
+  auto id = restable->at(row, col);
+  return id.getDatatype() == Datatype::GeoPoint
+             ? std::optional{id.getGeoPoint()}
+             : std::nullopt;
+};
+
+// ____________________________________________________________________________
+Id SpatialJoinAlgorithms::computeDist(const IdTable* idTableLeft,
+                                      const IdTable* idTableRight,
+                                      size_t rowLeft, size_t rowRight,
+                                      ColumnIndex leftPointCol,
+                                      ColumnIndex rightPointCol) const {
+  auto point1 = getPoint(idTableLeft, rowLeft, leftPointCol);
+  auto point2 = getPoint(idTableRight, rowRight, rightPointCol);
+  if (!point1.has_value() || !point2.has_value()) {
+    return Id::makeUndefined();
+  }
+  return Id::makeFromInt(static_cast<long long>(
+      ad_utility::detail::wktDistImpl(point1.value(), point2.value()) * 1000));
+}
+
+// ____________________________________________________________________________
+void SpatialJoinAlgorithms::addResultTableEntry(IdTable* result,
+                                                const IdTable* idTableLeft,
+                                                const IdTable* idTableRight,
+                                                size_t rowLeft, size_t rowRight,
+                                                Id distance) const {
+  // this lambda function copies elements from copyFrom
+  // into the table res. It copies them into the row
+  // rowIndRes and column column colIndRes. It returns the column number
+  // until which elements were copied
+  auto addColumns = [](IdTable* res, const IdTable* copyFrom, size_t rowIndRes,
+                       size_t colIndRes, size_t rowIndCopy) {
+    size_t col = 0;
+    while (col < copyFrom->numColumns()) {
+      res->at(rowIndRes, colIndRes) = (*copyFrom).at(rowIndCopy, col);
+      colIndRes += 1;
+      col += 1;
+    }
+    return colIndRes;
+  };
+
+  auto resrow = result->numRows();
+  result->emplace_back();
+  // add columns to result table
+  size_t rescol = 0;
+  rescol = addColumns(result, idTableLeft, resrow, rescol, rowLeft);
+  rescol = addColumns(result, idTableRight, resrow, rescol, rowRight);
+
+  if (addDistToResult_) {
+    result->at(resrow, rescol) = distance;
+    // rescol isn't used after that in this function, but future updates,
+    // which add additional columns, would need to remember to increase
+    // rescol at this place otherwise. If they forget to do this, the
+    // distance column will be overwritten, the variableToColumnMap will
+    // not work and so on
+    // rescol += 1;
+  }
+}
+
+// ____________________________________________________________________________
+Result SpatialJoinAlgorithms::BaselineAlgorithm() {
+  const auto [idTableLeft, resultLeft, idTableRight, resultRight, leftJoinCol,
+              rightJoinCol, numColumns, maxDist, maxResults] = params_;
+  IdTable result{numColumns, qec_->getAllocator()};
+
+  // cartesian product between the two tables, pairs are restricted according to
+  // `maxDistance_` and `maxResults_`
+  for (size_t rowLeft = 0; rowLeft < idTableLeft->size(); rowLeft++) {
+    // This priority queue stores the intermediate best results if `maxResults_`
+    // is used. Each intermediate result is stored as a pair of its `rowRight`
+    // and distance. Since the queue will hold at most `maxResults_ + 1` items,
+    // it is not a memory concern.
+    auto compare = [](std::pair<size_t, int64_t> a,
+                      std::pair<size_t, int64_t> b) {
+      return a.second < b.second;
+    };
+    std::priority_queue<std::pair<size_t, int64_t>,
+                        std::vector<std::pair<size_t, int64_t>>,
+                        decltype(compare)>
+        intermediate(compare);
+
+    // Inner loop of cartesian product
+    for (size_t rowRight = 0; rowRight < idTableRight->size(); rowRight++) {
+      Id dist = computeDist(idTableLeft, idTableRight, rowLeft, rowRight,
+                            leftJoinCol, rightJoinCol);
+
+      // Ensure `maxDist_` constraint
+      if (dist.getDatatype() != Datatype::Int ||
+          (maxDist.has_value() &&
+           dist.getInt() > static_cast<int64_t>(maxDist.value()))) {
+        continue;
+      }
+
+      // If there is no `maxResults_` we can add the result row immediately
+      if (!maxResults.has_value()) {
+        addResultTableEntry(&result, idTableLeft, idTableRight, rowLeft,
+                            rowRight, dist);
+        continue;
+      }
+
+      // Ensure `maxResults_` constraint using priority queue
+      intermediate.push(std::pair{rowRight, dist.getInt()});
+      // Too many results? Drop the worst one
+      if (intermediate.size() > maxResults.value()) {
+        intermediate.pop();
+      }
+    }
+
+    // If we are using the priority queue, we didn't add the results in the
+    // inner loop, so we do it now.
+    if (maxResults.has_value()) {
+      size_t numResults = intermediate.size();
+      for (size_t item = 0; item < numResults; item++) {
+        // Get and remove largest item from priority queue
+        auto [rowRight, dist] = intermediate.top();
+        intermediate.pop();
+
+        addResultTableEntry(&result, idTableLeft, idTableRight, rowLeft,
+                            rowRight, Id::makeFromInt(dist));
+      }
+    }
+  }
+  return Result(std::move(result), std::vector<ColumnIndex>{},
+                Result::getMergedLocalVocab(*resultLeft, *resultRight));
+}
+
+// ____________________________________________________________________________
+Result SpatialJoinAlgorithms::S2geometryAlgorithm() {
+  const auto [idTableLeft, resultLeft, idTableRight, resultRight, leftJoinCol,
+              rightJoinCol, numColumns, maxDist, maxResults] = params_;
+  IdTable result{numColumns, qec_->getAllocator()};
+
+  // Helper function to convert `GeoPoint` to `S2Point`
+  auto constexpr toS2Point = [](const GeoPoint& p) {
+    auto lat = p.getLat();
+    auto lng = p.getLng();
+    auto latlng = S2LatLng::FromDegrees(lat, lng);
+    return S2Point{latlng};
+  };
+
+  S2PointIndex<size_t> s2index;
+
+  // Optimization: If we only search by maximum distance, the operation is
+  // symmetric, so the larger table can be used for the index
+  bool indexOfRight =
+      (maxResults.has_value() || (idTableLeft->size() > idTableRight->size()));
+  auto indexTable = indexOfRight ? idTableRight : idTableLeft;
+  auto indexJoinCol = indexOfRight ? rightJoinCol : leftJoinCol;
+
+  // Populate the index
+  for (size_t row = 0; row < indexTable->size(); row++) {
+    auto p = getPoint(indexTable, row, indexJoinCol);
+    if (p.has_value()) {
+      s2index.Add(toS2Point(p.value()), row);
+    }
+  }
+
+  // Performs a nearest neighbor search on the index and returns the closest
+  // points that satisfy the criteria given by `maxDist_` and `maxResults_`.
+
+  // Construct a query object with the given constraints
+  auto s2query = S2ClosestPointQuery<size_t>{&s2index};
+
+  if (maxResults.has_value()) {
+    s2query.mutable_options()->set_max_results(
+        static_cast<int>(maxResults.value()));
+  }
+  if (maxDist.has_value()) {
+    s2query.mutable_options()->set_inclusive_max_distance(S2Earth::ToAngle(
+        util::units::Meters(static_cast<float>(maxDist.value()))));
+  }
+
+  auto searchTable = indexOfRight ? idTableLeft : idTableRight;
+  auto searchJoinCol = indexOfRight ? leftJoinCol : rightJoinCol;
+
+  // Use the index to lookup the points of the other table
+  for (size_t searchRow = 0; searchRow < searchTable->size(); searchRow++) {
+    auto p = getPoint(searchTable, searchRow, searchJoinCol);
+    if (!p.has_value()) {
+      continue;
+    }
+    auto s2target =
+        S2ClosestPointQuery<size_t>::PointTarget{toS2Point(p.value())};
+
+    for (const auto& neighbor : s2query.FindClosestPoints(&s2target)) {
+      // In this loop we only receive points that already satisfy the given
+      // criteria
+      auto indexRow = neighbor.data();
+      auto dist = static_cast<int64_t>(S2Earth::ToMeters(neighbor.distance()));
+
+      auto rowLeft = indexOfRight ? searchRow : indexRow;
+      auto rowRight = indexOfRight ? indexRow : searchRow;
+      addResultTableEntry(&result, idTableLeft, idTableRight, rowLeft, rowRight,
+                          Id::makeFromInt(dist));
+    }
+  }
+
+  return Result(std::move(result), std::vector<ColumnIndex>{},
+                Result::getMergedLocalVocab(*resultLeft, *resultRight));
+}
