+++ test/engine/CMakeLists.txt
@@ -9,3 +9,4 @@ addLinkAndDiscoverTest(GroupByHashMapOptimizationTest)
 addLinkAndDiscoverTest(LazyGroupByTest engine)
 addLinkAndDiscoverTest(CountConnectedSubgraphsTest)
 addLinkAndDiscoverTest(BindTest engine)
+addLinkAndDiscoverTest(SpatialJoinAlgorithmsTest engine)
